Backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
openai==1.99.2
requests==2.32.4
python-jose==3.5.0
pydantic==2.11.7
PyMuPDF==1.26.3
beautifulsoup4==4.13.4
lxml==6.0.0
striprtf==0.0.26
gunicorn==21.2.0
boto3==1.34.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pgvector==0.2.3
